Fix compilation warnings, related to circular dependency Owned/Gltf.
